runtime: Move low-level config loading into internal package
This allows it to be reused by commands other than opa run.
