+++ runtime/runtime_test.go
@@ -7,17 +7,13 @@ package runtime
 import (
 	"bytes"
 	"context"
-	"fmt"
 	"io/ioutil"
 	"os"
 	"path"
-	"path/filepath"
 	"reflect"
 	"strings"
 	"testing"
 	"time"
-
-	"github.com/ghodss/yaml"
 
 	"github.com/open-policy-agent/opa/ast"
 	"github.com/open-policy-agent/opa/storage"
@@ -237,153 +233,4 @@ func testRuntimeProcessWatchEventPolicyError(t *testing.T, asBundle bool) {
 		}
 
 	})
-}
-
-func TestLoadConfigWithParamOverride(t *testing.T) {
-	fs := map[string]string{"/some/config.yaml": `
-services:
-  acmecorp:
-    url: https://example.com/control-plane-api/v1
-
-discovery:
-  name: /example/discovery
-  prefix: configuration
-`}
-
-	test.WithTempFS(fs, func(rootDir string) {
-		params := NewParams()
-		params.ConfigFile = filepath.Join(rootDir, "/some/config.yaml")
-		params.ConfigOverrides = []string{"services.acmecorp.credentials.bearer.token=bGFza2RqZmxha3NkamZsa2Fqc2Rsa2ZqYWtsc2RqZmtramRmYWxkc2tm"}
-
-		configBytes, err := loadConfig(params)
-		if err != nil {
-			t.Errorf("unexpected error loading config: " + err.Error())
-		}
-
-		config := map[string]interface{}{}
-		err = yaml.Unmarshal(configBytes, &config)
-		if err != nil {
-			t.Errorf("unexpected error unmarshalling config")
-		}
-
-		expected := map[string]interface{}{
-			"services": map[string]interface{}{
-				"acmecorp": map[string]interface{}{
-					"url": "https://example.com/control-plane-api/v1",
-					"credentials": map[string]interface{}{
-						"bearer": map[string]interface{}{
-							"token": "bGFza2RqZmxha3NkamZsa2Fqc2Rsa2ZqYWtsc2RqZmtramRmYWxkc2tm",
-						},
-					},
-				},
-			},
-			"discovery": map[string]interface{}{
-				"name":   "/example/discovery",
-				"prefix": "configuration",
-			},
-		}
-
-		if !reflect.DeepEqual(config, expected) {
-			t.Errorf("config does not match expected:\n\nExpected: %+v\nActual: %+v", expected, config)
-		}
-	})
-}
-
-func TestLoadConfigWithFileOverride(t *testing.T) {
-	fs := map[string]string{
-		"/some/config.yaml": `
-services:
-  acmecorp:
-    url: https://example.com/control-plane-api/v1
-    credentials:
-      bearer:
-        token: "XXXXXXXXXX"
-
-discovery:
-  name: /example/discovery
-  prefix: configuration
-`,
-		"/some/secret.txt": "bGFza2RqZmxha3NkamZsa2Fqc2Rsa2ZqYWtsc2RqZmtramRmYWxkc2tm",
-	}
-
-	test.WithTempFS(fs, func(rootDir string) {
-		params := NewParams()
-		params.ConfigFile = filepath.Join(rootDir, "/some/config.yaml")
-		secretFile := filepath.Join(rootDir, "/some/secret.txt")
-		params.ConfigOverrideFiles = []string{fmt.Sprintf("services.acmecorp.credentials.bearer.token=%s", secretFile)}
-
-		configBytes, err := loadConfig(params)
-		if err != nil {
-			t.Errorf("unexpected error loading config: " + err.Error())
-		}
-
-		config := map[string]interface{}{}
-		err = yaml.Unmarshal(configBytes, &config)
-		if err != nil {
-			t.Errorf("unexpected error unmarshalling config")
-		}
-
-		expected := map[string]interface{}{
-			"services": map[string]interface{}{
-				"acmecorp": map[string]interface{}{
-					"url": "https://example.com/control-plane-api/v1",
-					"credentials": map[string]interface{}{
-						"bearer": map[string]interface{}{
-							"token": "bGFza2RqZmxha3NkamZsa2Fqc2Rsa2ZqYWtsc2RqZmtramRmYWxkc2tm",
-						},
-					},
-				},
-			},
-			"discovery": map[string]interface{}{
-				"name":   "/example/discovery",
-				"prefix": "configuration",
-			},
-		}
-
-		if !reflect.DeepEqual(config, expected) {
-			t.Errorf("config does not match expected:\n\nExpected: %+v\nActual: %+v", expected, config)
-		}
-	})
-}
-
-func TestLoadConfigWithParamOverrideNoConfigFile(t *testing.T) {
-	params := NewParams()
-	params.ConfigOverrides = []string{
-		"services.acmecorp.url=https://example.com/control-plane-api/v1",
-		"services.acmecorp.credentials.bearer.token=bGFza2RqZmxha3NkamZsa2Fqc2Rsa2ZqYWtsc2RqZmtramRmYWxkc2tm",
-		"discovery.name=/example/discovery",
-		"discovery.prefix=configuration",
-	}
-
-	configBytes, err := loadConfig(params)
-	if err != nil {
-		t.Errorf("unexpected error loading config: " + err.Error())
-	}
-
-	config := map[string]interface{}{}
-	err = yaml.Unmarshal(configBytes, &config)
-	if err != nil {
-		t.Errorf("unexpected error unmarshalling config")
-	}
-
-	expected := map[string]interface{}{
-		"services": map[string]interface{}{
-			"acmecorp": map[string]interface{}{
-				"url": "https://example.com/control-plane-api/v1",
-				"credentials": map[string]interface{}{
-					"bearer": map[string]interface{}{
-						"token": "bGFza2RqZmxha3NkamZsa2Fqc2Rsa2ZqYWtsc2RqZmtramRmYWxkc2tm",
-					},
-				},
-			},
-		},
-		"discovery": map[string]interface{}{
-			"name":   "/example/discovery",
-			"prefix": "configuration",
-		},
-	}
-
-	if !reflect.DeepEqual(config, expected) {
-		t.Errorf("config does not match expected:\n\nExpected: %+v\nActual: %+v", expected, config)
-	}
 }
